Move `bind` info to `Frame` methods.
